refactor: extract DownloadTask from ApplicationController
